версия giswsdl 11.7.0.10
Former-commit-id: d7c24b3d28b788f0febdd922917d66c198dc9342

+++ src/main/java/ru/progmatik/java/pregis/connectiondb/grad/account/AccountGRADDAO.java
@@ -290,7 +290,7 @@ public final class AccountGRADDAO {
 // debug                    answerProcessing.sendMessageToClient(basicInformation.getNumberLS() + " found!");
                 ImportAccountRequest.Account account = new ImportAccountRequest.Account();
 //                    account.setCreationDate(OtherFormat.getDateNow()); // может без даты можно? добавил при отправки нового счета
-                account.setLivingPersonsNumber((byte) basicInformation.getAmountLiving());
+                account.setLivingPersonsNumber(basicInformation.getAmountLiving());
                 account.setTotalSquare(new BigDecimal(basicInformation.getTotalArea()).setScale(2, BigDecimal.ROUND_DOWN));
                 account.setResidentialSquare(new BigDecimal(basicInformation.getLivingSpace()).setScale(2, BigDecimal.ROUND_DOWN));
                 if (basicInformation.getHeadtedArea() > 0.0) {
